requirements.txt
# python requirements
openmesh>=1.1.2
opencv-python>=3.4.2.17
PyOpenGL>=3.1.0
numpy
scipy
matplotlib
Pillow
scikit-learn
scikit-image
cffi
tqdm
configparser
absl-py
ipdb
git+https://github.com/yig/PySPQR.git
